tests/test_main.py
import pytest
from doc_oct import main
from unittest.mock import patch, MagicMock

# Mock the command-line arguments
@pytest.fixture
def mock_args():
    with patch('argparse.ArgumentParser.parse_args') as mock_parse_args:
        yield mock_parse_args

# Mock the other functions called within main()
@pytest.fixture
def mock_functions():
    with patch('doc_oct.main.prompt_for_confirmation') as mock_prompt, \
         patch('doc_oct.main.scan_directory') as mock_scan, \
         patch('doc_oct.main.backup_file') as mock_backup, \
         patch('doc_oct.main.DocstringParser') as mock_parser, \
         patch('doc_oct.main.DocstringWriter') as mock_writer, \
         patch('doc_oct.main.Logger') as mock_logger:
        mock_prompt.return_value = True
        mock_scan.return_value = ['file1.py', 'file2.py']
        mock_parser.get_docstring = MagicMock(return_value=None)
        yield mock_prompt, mock_scan, mock_backup, mock_parser, mock_writer, mock_logger

def test_main_functionality(mock_args, mock_functions):
    """
    Test the main function with mocked dependencies.
    """
    # Setup mock arguments
    mock_args.return_value = MagicMock(directory='test_dir', dry_run=False, interactive=False)

    # Call main()
    main.main('test_dir', False, False)

    # Assert that the necessary functions were called
    mock_functions[1].assert_called_with('test_dir')  # scan_directory
    mock_functions[3].get_docstring.assert_any_call('file1.py')  # DocstringParser.get_docstring
    mock_functions[4].write_docstring.assert_any_call('file1.py', '"""\nYour docstring here.\n"""')  # DocstringWriter.write_docstring
    # Additional assertions can be added as needed

def test_main_dry_run(mock_args, mock_functions):
    """
    Test the main function in dry-run mode.
    """
    mock_args.return_value = MagicMock(directory='test_dir', dry_run=True, interactive=False)
    main.main('test_dir', True, False)
    # Similar assertions can be made for dry-run mode
